feat(zimbra.dashboard): add dashboard manager-components
ref:MANAGER-13039
